Added item pouch

+++ src/main/java/com/github/thomashooks/notenoughmachines/integration/config/CommonConfigs.java
@@ -11,8 +11,8 @@ public class CommonConfigs
     public static final ForgeConfigSpec.ConfigValue<Float> HIGH_SPEED_RAIL_MAX_SPEED_WATERLOGGED;
     public static final ForgeConfigSpec.ConfigValue<Float> HIGH_SPEED_RAIL_MAX_SPEED_MINECART_FURNACE;
     public static final ForgeConfigSpec.ConfigValue<Float> HIGH_SPEED_RAIL_MAX_SPEED_MINECART_FURNACE_WATERLOGGED;
-    public static final ForgeConfigSpec.ConfigValue<Double> HIGH_SPEED_POWERED_RAIL_BOOST_FACTOR;
-    public static final ForgeConfigSpec.ConfigValue<Double> HIGH_SPEED_POWERED_RAIL_LAUNCH_FACTOR;
+    public static final ForgeConfigSpec.ConfigValue<Float> HIGH_SPEED_POWERED_RAIL_BOOST_FACTOR;
+    public static final ForgeConfigSpec.ConfigValue<Float> HIGH_SPEED_POWERED_RAIL_LAUNCH_FACTOR;
 
     static
     {
@@ -27,9 +27,9 @@ public class CommonConfigs
         HIGH_SPEED_RAIL_MAX_SPEED_MINECART_FURNACE_WATERLOGGED = BUILDER.comment("The maximum speed of a furnace minecart while it's on a waterlogged high-speed rail")
                 .define("Furnace Minecart's Max Speed on a waterlogged High-Speed Rail", 0.3F);
         HIGH_SPEED_POWERED_RAIL_BOOST_FACTOR = BUILDER.comment("The boost factor of a minecart while it's on a high-speed powered rail")
-                .define("High-Speed Powered Rail's Boost Factor", 0.08D);
+                .define("High-Speed Powered Rail's Boost Factor", 0.08F);
         HIGH_SPEED_POWERED_RAIL_LAUNCH_FACTOR = BUILDER.comment("The launch factor of a minecart while it's on a high-speed powered rail")
-                .define("High-Speed Powered Rail's Launch Factor", 0.04D);
+                .define("High-Speed Powered Rail's Launch Factor", 0.04F);
 
         BUILDER.pop();
         SPEC = BUILDER.build();
